fix: pass getToken form frontend to backend

+++ src/lib/apiService.ts
@@ -1,10 +1,25 @@
 import { getApiUrl } from "@/utils/ApiHelper";
+
 const apiUrl = getApiUrl();
-const token = "your_jwt_bearer_token_here";
 
 const apiService = {
+  async fetchToken(): Promise<string | null> {
+    try {
+      const response = await fetch("/api/auth");
+      if (!response.ok) {
+        throw new Error("Failed to fetch token");
+      }
+      const data = await response.json();
+      return data.token;
+    } catch (error) {
+      console.error("Error fetching token:", error);
+      return null;
+    }
+  },
+
   async get<T>(endpoint: string): Promise<T> {
     try {
+      const token = await this.fetchToken();
       const response = await fetch(`${apiUrl}${endpoint}`, {
         headers: {
           Authorization: `Bearer ${token}`,
@@ -13,7 +28,9 @@ const apiService = {
       if (!response.ok) {
         throw new Error(`HTTP error! status: ${response.status}`);
       }
-      return await response.json();
+      const result = (await response.json()) as T;
+      console.log("result", result);
+      return result;
     } catch (error) {
       console.error(error);
       throw error;
@@ -21,8 +38,11 @@ const apiService = {
   },
 
   async post<T>(endpoint: string, data: any): Promise<T> {
-    console.log(`${apiUrl}${endpoint}`);
     try {
+      const token = await this.fetchToken();
+      if (!token) {
+        throw new Error("Token is not available");
+      }
       const response = await fetch(`${apiUrl}${endpoint}`, {
         method: "POST",
         headers: {
@@ -34,7 +54,9 @@ const apiService = {
       if (!response.ok) {
         throw new Error(`HTTP error! status: ${response.status}`);
       }
-      return await response.json();
+      const result = (await response.json()) as T;
+      console.log("result", result);
+      return result;
     } catch (error) {
       console.error(error);
       throw error;
@@ -43,6 +65,10 @@ const apiService = {
 
   async put<T>(endpoint: string, data: any): Promise<T> {
     try {
+      const token = await this.fetchToken();
+      if (!token) {
+        throw new Error("Token is not available");
+      }
       const response = await fetch(`${apiUrl}${endpoint}`, {
         method: "PUT",
         headers: {
@@ -63,6 +89,10 @@ const apiService = {
 
   async delete<T>(endpoint: string): Promise<T> {
     try {
+      const token = await this.fetchToken();
+      if (!token) {
+        throw new Error("Token is not available");
+      }
       const response = await fetch(`${apiUrl}${endpoint}`, {
         method: "DELETE",
         headers: {
